package no longer depends on sled-agent (#2468)
`package` now depends on `sled-hardware`, which is a smaller interface

- Pulls `cleanup_networking_resources` and friends into
`sled-hardware/cleanup`
- Pulls `bootstrap/agent`'s `delete_omicron_vnics` into
`sled-hardware/cleanup`
- Pulls `sled-agent/common/underlay` into `sled-hardware/underlay`

Fixes https://github.com/oxidecomputer/omicron/issues/2458
