+++ package/Cargo.toml
@@ -13,7 +13,6 @@ hex.workspace = true
 illumos-utils.workspace = true
 indicatif.workspace = true
 omicron-common.workspace = true
-omicron-sled-agent.workspace = true
 omicron-zone-package.workspace = true
 petgraph.workspace = true
 rayon.workspace = true
@@ -21,6 +20,7 @@ reqwest = { workspace = true, features = [ "rustls-tls" ] }
 ring.workspace = true
 serde.workspace = true
 serde_derive.workspace = true
+sled-hardware.workspace = true
 slog.workspace = true
 slog-async.workspace = true
 slog-term.workspace = true
