+++ package/src/bin/omicron-package.rs
@@ -10,13 +10,13 @@ use futures::stream::{self, StreamExt, TryStreamExt};
 use illumos_utils::{zfs, zone, zpool};
 use indicatif::{MultiProgress, ProgressBar, ProgressStyle};
 use omicron_package::{parse, BuildCommand, DeployCommand};
-use omicron_sled_agent::cleanup_networking_resources;
 use omicron_zone_package::config::Config as PackageConfig;
 use omicron_zone_package::package::{Package, PackageOutput, PackageSource};
 use omicron_zone_package::progress::Progress;
 use omicron_zone_package::target::Target;
 use rayon::prelude::*;
 use ring::digest::{Context as DigestContext, Digest, SHA256};
+use sled_hardware::cleanup::cleanup_networking_resources;
 use slog::debug;
 use slog::o;
 use slog::Drain;
