+++ sled-agent/src/bootstrap/server.rs
@@ -13,11 +13,11 @@ use super::trust_quorum::ShareDistribution;
 use super::views::Response;
 use super::views::ResponseEnvelope;
 use crate::bootstrap::maghemite;
-use crate::common::underlay;
 use crate::config::Config as SledConfig;
 use crate::sp::AsyncReadWrite;
 use crate::sp::SpHandle;
 use crate::sp::SprocketsRole;
+use sled_hardware::underlay;
 use slog::Drain;
 use slog::Logger;
 use std::sync::Arc;
