+++ sled-agent/src/sled_agent.rs
@@ -18,7 +18,6 @@ use crate::services::{self, ServiceManager};
 use crate::storage_manager::StorageManager;
 use crate::updates::{ConfigUpdates, UpdateManager};
 use dropshot::HttpError;
-use illumos_utils::zone::IPADM;
 use illumos_utils::{execute, PFEXEC};
 use omicron_common::address::{
     get_sled_address, get_switch_zone_address, Ipv6Subnet, SLED_PREFIX,
@@ -30,11 +29,11 @@ use omicron_common::api::{
 use omicron_common::backoff::{
     retry_notify, retry_policy_internal_service_aggressive, BackoffError,
 };
+use sled_hardware::underlay;
 use sled_hardware::HardwareManager;
 use slog::Logger;
 use std::net::{Ipv6Addr, SocketAddrV6};
 use std::path::PathBuf;
-use std::process::Command;
 use std::sync::Arc;
 use uuid::Uuid;
 
@@ -65,7 +64,7 @@ pub enum Error {
     DeleteAddress(#[from] illumos_utils::ExecutionError),
 
     #[error("Failed to operate on underlay device: {0}")]
-    Underlay(#[from] crate::common::underlay::Error),
+    Underlay(#[from] underlay::Error),
 
     #[error(transparent)]
     Services(#[from] crate::services::Error),
@@ -224,7 +223,7 @@ impl SledAgent {
         .map_err(|err| Error::SledSubnet { err })?;
 
         // Initialize the xde kernel driver with the underlay devices.
-        let underlay_nics = crate::common::underlay::find_nics()?;
+        let underlay_nics = underlay::find_nics()?;
         illumos_utils::opte::initialize_xde_driver(&log, &underlay_nics)?;
 
         // Ipv6 forwarding must be enabled to route traffic between zones.
@@ -595,85 +594,4 @@ impl SledAgent {
             .await
             .map_err(Error::from)
     }
-}
-
-// Delete all underlay addresses created directly over the etherstub VNICs used
-// for inter-zone communications.
-fn delete_etherstub_addresses(log: &Logger) -> Result<(), Error> {
-    let underlay_prefix =
-        format!("{}/", illumos_utils::dladm::UNDERLAY_ETHERSTUB_VNIC_NAME);
-    let bootstrap_prefix =
-        format!("{}/", illumos_utils::dladm::BOOTSTRAP_ETHERSTUB_VNIC_NAME);
-    delete_addresses_matching_prefixes(log, &[underlay_prefix])?;
-    delete_addresses_matching_prefixes(log, &[bootstrap_prefix])
-}
-
-fn delete_underlay_addresses(log: &Logger) -> Result<(), Error> {
-    use illumos_utils::dladm::VnicSource;
-    let prefixes = crate::common::underlay::find_chelsio_links()?
-        .into_iter()
-        .map(|link| format!("{}/", link.name()))
-        .collect::<Vec<_>>();
-    delete_addresses_matching_prefixes(log, &prefixes)
-}
-
-fn delete_addresses_matching_prefixes(
-    log: &Logger,
-    prefixes: &[String],
-) -> Result<(), Error> {
-    use std::io::BufRead;
-    let mut cmd = Command::new(PFEXEC);
-    let cmd = cmd.args(&[IPADM, "show-addr", "-p", "-o", "ADDROBJ"]);
-    let output = execute(cmd)?;
-
-    // `ipadm show-addr` can return multiple addresses with the same name, but
-    // multiple values. Collecting to a set ensures that only a single name is
-    // used.
-    let addrobjs = output
-        .stdout
-        .lines()
-        .flatten()
-        .collect::<std::collections::HashSet<_>>();
-
-    for addrobj in addrobjs {
-        if prefixes.iter().any(|prefix| addrobj.starts_with(prefix)) {
-            warn!(
-                log,
-                "Deleting existing Omicron IP address";
-                "addrobj" => addrobj.as_str(),
-            );
-            let mut cmd = Command::new(PFEXEC);
-            let cmd = cmd.args(&[IPADM, "delete-addr", addrobj.as_str()]);
-            execute(cmd)?;
-        }
-    }
-    Ok(())
-}
-
-// Delete the etherstub and underlay VNIC used for interzone communication
-fn delete_etherstub(log: &Logger) -> Result<(), Error> {
-    use illumos_utils::dladm::BOOTSTRAP_ETHERSTUB_NAME;
-    use illumos_utils::dladm::BOOTSTRAP_ETHERSTUB_VNIC_NAME;
-    use illumos_utils::dladm::UNDERLAY_ETHERSTUB_NAME;
-    use illumos_utils::dladm::UNDERLAY_ETHERSTUB_VNIC_NAME;
-    warn!(log, "Deleting Omicron underlay VNIC"; "vnic_name" => UNDERLAY_ETHERSTUB_VNIC_NAME);
-    Dladm::delete_etherstub_vnic(UNDERLAY_ETHERSTUB_VNIC_NAME)?;
-    warn!(log, "Deleting Omicron underlay etherstub"; "stub_name" => UNDERLAY_ETHERSTUB_NAME);
-    Dladm::delete_etherstub(UNDERLAY_ETHERSTUB_NAME)?;
-    warn!(log, "Deleting Omicron bootstrap VNIC"; "vnic_name" => BOOTSTRAP_ETHERSTUB_VNIC_NAME);
-    Dladm::delete_etherstub_vnic(BOOTSTRAP_ETHERSTUB_VNIC_NAME)?;
-    warn!(log, "Deleting Omicron bootstrap etherstub"; "stub_name" => BOOTSTRAP_ETHERSTUB_NAME);
-    Dladm::delete_etherstub(BOOTSTRAP_ETHERSTUB_NAME)?;
-    Ok(())
-}
-
-/// Delete all networking resources installed by the sled agent, in the global
-/// zone.
-pub async fn cleanup_networking_resources(log: &Logger) -> Result<(), Error> {
-    delete_etherstub_addresses(log)?;
-    delete_underlay_addresses(log)?;
-    crate::bootstrap::agent::delete_omicron_vnics(log).await?;
-    delete_etherstub(log)?;
-    illumos_utils::opte::delete_all_xde_devices(log)?;
-    Ok(())
 }
