+++ sled-hardware/Cargo.toml
@@ -8,6 +8,7 @@ license = "MPL-2.0"
 [dependencies]
 anyhow.workspace = true
 cfg-if.workspace = true
+futures.workspace = true
 illumos-utils.workspace = true
 nexus-client.workspace = true
 slog.workspace = true
